Create a SourceFooter generic component and use it for PJ
