+++ src/scss/includes/components/sourceFooter.scss
@@ -1,0 +1,24 @@
+
+.sourceFooter {
+  width: 100%;
+  height: 32px;
+  border-top: 1px solid $grey-light;
+  line-height: 32px;
+  font-size: 12px;
+  padding: 0 14px;
+  color: $grey-semi-darkness;
+}
+
+@media (max-width: 640px) {
+  .sourceFooter {
+    text-align: center;
+    position: fixed;
+    bottom: 0;
+    height: 36px;
+    line-height: 36px;
+    font-size: 14px;
+    border-radius: 0;
+    background-color: white;
+    z-index: 3;
+  }
+}
