+++ src/scss/includes/no_ui.scss
@@ -16,7 +16,7 @@
     visibility: hidden;
   }
 
-  .category__panel__pj {
+  .sourceFooter {
     visibility: hidden;
   }
