+++ src/scss/includes/panels/categories.scss
@@ -9,16 +9,6 @@
   }
 }
 
-.category__panel__pj {
-  width: 100%;
-  height: 32px;
-  border-top: 1px solid $grey-light;
-  line-height: 32px;
-  font-size: 12px;
-  padding: 0 14px;
-  color: $grey-semi-darkness;
-}
-
 .category__panel__items {
   animation: appear 600ms forwards;
 }
@@ -63,18 +53,6 @@
     &.panel--pj {
       padding: 0 0 36px;
     }
-  }
-
-  .category__panel__pj {
-    text-align: center;
-    position: fixed;
-    bottom: 0;
-    height: 36px;
-    line-height: 36px;
-    font-size: 14px;
-    border-radius: 0;
-    background-color: white;
-    z-index: 3;
   }
 
   .category__panel__error {
