+++ src/scss/includes/ui_components.scss
@@ -15,3 +15,4 @@
 @import "./components/floatingButton";
 @import "./components/floatingItems";
 @import "./components/backToQwantButton";
+@import "./components/sourceFooter";
